address reviewer suggestion
